Fixed naming bugs (again) by moving subsystems to 'subs' folder. Do not delete empty Subsystems folder or rename subs folder. Also modified algae speed
