+++ src/main/java/frc/robot/commands/AutoAlignToReefRight.java
@@ -9,8 +9,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.LimelightHelpers;
 import frc.robot.Constants.LimelightConstants;
 import frc.robot.Constants.SwerveDrivebaseConstants;
-import frc.robot.Subsystems.Swerve;
-
+import frc.robot.subs.Swerve;
 
 public class AutoAlignToReefRight extends Command {
     private final Swerve m_db;
