+++ src/main/java/frc/robot/commands/LiftAlgae.java
@@ -8,8 +8,8 @@ import edu.wpi.first.math.controller.ElevatorFeedforward;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.Constants.ElevatorConstants;
-import frc.robot.Subsystems.Algae;
-import frc.robot.subsystems.Elevator;
+import frc.robot.subs.Algae;
+import frc.robot.subs.Elevator;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
 public class LiftAlgae extends Command {
