+++ src/main/java/frc/robot/subs/Algae.java
@@ -1,4 +1,4 @@
-package frc.robot.Subsystems;
+package frc.robot.subs;
 
 import com.revrobotics.RelativeEncoder;
 import com.revrobotics.spark.SparkBase.PersistMode;
