+++ src/main/java/frc/robot/subs/Swerve.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.Subsystems;
+package frc.robot.subs;
 
 import java.util.function.Supplier;
 
@@ -217,7 +217,7 @@ public class Swerve extends SubsystemBase {
     SmartDashboard.putNumber("TX", LimelightHelpers.getTX(LimelightConstants.kName));
     SmartDashboard.putNumber("TY", LimelightHelpers.getTY(LimelightConstants.kName));
     SmartDashboard.putNumber("Straight Distance", this.getStraightDistance());
-    SmartDashboard.putNumber("Offset", this.calculateDistanceFromAprilTag());
+    // SmartDashboard.putNumber("Offset", this.calculateDistanceFromAprilTag());
     tunePIDSmartDashboard();
   }
